requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
boto3
python-dotenv
requests
pandas
numpy
tqdm
# AI/ML
transformers
torch
torchaudio
librosa
faiss-cpu
google-genai
